Revert "Merge branch 'main' into bugfix/operator2_resources_errs"
This reverts commit 34ea62ac2ec2d68fd8529a00653c9f4d51e0a04b, reversing
changes made to 8c019b87fb0034e8feba65b7f860f82af369b0d7.
