+++ pennylane/labs/phox/mmd_loss.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""MMD loss utilities for tcdq."""
+"""MMD loss utilities for Phox."""
 
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass
