+++ pennylane/ops/op_math/linear_combination.py
@@ -22,9 +22,8 @@ import numbers
 from copy import copy
 
 import pennylane as qp
-from pennylane.core.operator import Operator, Operator2
+from pennylane.core.operator import Operator
 from pennylane.core.queuing import AnnotatedQueue, QueuingManager
-from pennylane.pytrees import flatten, unflatten
 
 from .sprod import SProd
 from .sum import Sum
@@ -119,21 +118,7 @@ class LinearCombination(Sum):
     # pylint: disable=arguments-differ
     @classmethod
     def _primitive_bind_call(cls, coeffs, observables, _pauli_rep=None, **kwargs):
-        leaves, structure = flatten(observables, is_leaf=lambda x: isinstance(x, Operator))
-
-        new_leaves = []
-        for leaf in leaves:
-            if isinstance(leaf, Operator2):
-                if leaf.tracer is None:
-                    # pylint: disable-next=protected-access
-                    leaf._bind_primitive()
-                new_leaves.append(leaf.tracer)
-            else:
-                new_leaves.append(leaf)
-
-        new_observables = unflatten(new_leaves, structure)
-
-        return cls._primitive.bind(*coeffs, *new_observables, **kwargs, n_obs=len(observables))
+        return cls._primitive.bind(*coeffs, *observables, **kwargs, n_obs=len(observables))
 
     def __init__(
         self,
