+++ pennylane/ops/op_math/symbolicop.py
@@ -22,12 +22,10 @@ from warnings import warn
 import numpy as np
 
 import pennylane as qp
-from pennylane import math as pl_math
-from pennylane.core.operator import Operator, Operator2
+from pennylane.core.operator import Operator
 from pennylane.core.operator.base import _UNSET_BATCH_SIZE  # tach-ignore
 from pennylane.core.queuing import QueuingManager
 from pennylane.exceptions import PennyLaneDeprecationWarning
-from pennylane.pytrees import flatten, unflatten
 
 from .composite import handle_recursion_error
 
@@ -55,21 +53,8 @@ class SymbolicOp(Operator):
 
     @classmethod
     def _primitive_bind_call(cls, *args, **kwargs):
-        leaves, structure = flatten((args, kwargs), is_leaf=lambda x: isinstance(x, Operator))
-
-        new_leaves = []
-        for leaf in leaves:
-            if isinstance(leaf, Operator2):
-                if leaf.tracer is None:
-                    # pylint: disable-next=protected-access
-                    leaf._bind_primitive()
-                new_leaves.append(leaf.tracer)
-            else:
-                new_leaves.append(leaf)
-
-        new_args, new_kwargs = unflatten(new_leaves, structure)
-        # needs to be overwritten because it doesnt take wires
-        return cls._primitive.bind(*new_args, **new_kwargs)
+        # has no wires, so doesn't need any wires processing
+        return cls._primitive.bind(*args, **kwargs)
 
     @handle_recursion_error
     def __copy__(self):
@@ -196,12 +181,12 @@ class ScalarSymbolicOp(SymbolicOp):
     def batch_size(self):
         if self._batch_size is _UNSET_BATCH_SIZE:
             base_batch_size = self.base.batch_size
-            if pl_math.ndim(self.scalar) == 0:
+            if qp.math.ndim(self.scalar) == 0:
                 # coeff is not batched
                 self._batch_size = base_batch_size
             else:
                 # coeff is batched
-                scalar_size = pl_math.size(self.scalar)
+                scalar_size = qp.math.size(self.scalar)
                 if base_batch_size is not None and base_batch_size != scalar_size:
                     raise ValueError(
                         "Broadcasting was attempted but the broadcasted dimensions "
@@ -273,35 +258,35 @@ class ScalarSymbolicOp(SymbolicOp):
         # compute base matrix
         base_matrix = self.base.matrix()
 
-        scalar_interface = pl_math.get_interface(self.scalar)
+        scalar_interface = qp.math.get_interface(self.scalar)
         scalar = self.scalar
         if scalar_interface == "torch":
             # otherwise get `RuntimeError: Can't call numpy() on Tensor that requires grad.`
-            base_matrix = pl_math.convert_like(base_matrix, self.scalar)
+            base_matrix = qp.math.convert_like(base_matrix, self.scalar)
         elif (
             scalar_interface == "tensorflow"
         ):  # pragma: no cover (TensorFlow tests were disabled during deprecation)
             # just cast everything to complex128. Otherwise we may have casting problems
             # where things get truncated like in SProd(tf.Variable(0.1), qp.X(0))
-            scalar = pl_math.cast(scalar, "complex128")
-            base_matrix = pl_math.cast(base_matrix, "complex128")
+            scalar = qp.math.cast(scalar, "complex128")
+            base_matrix = qp.math.cast(base_matrix, "complex128")
 
         # compute scalar operation on base matrix taking batching into account
-        scalar_size = pl_math.size(scalar)
+        scalar_size = qp.math.size(scalar)
         if scalar_size != 1:
             if scalar_size == self.base.batch_size:
                 # both base and scalar are broadcasted
-                mat = pl_math.stack(
+                mat = qp.math.stack(
                     [self._matrix(s, m) for s, m in zip(scalar, base_matrix, strict=True)]
                 )
             else:
                 # only scalar is broadcasted
-                mat = pl_math.stack([self._matrix(s, base_matrix) for s in scalar])
+                mat = qp.math.stack([self._matrix(s, base_matrix) for s in scalar])
         elif self.base.batch_size is not None:
             # only base is broadcasted
-            mat = pl_math.stack([self._matrix(scalar, ar2) for ar2 in base_matrix])
+            mat = qp.math.stack([self._matrix(scalar, ar2) for ar2 in base_matrix])
         else:
             # none are broadcasted
             mat = self._matrix(scalar, base_matrix)
 
-        return pl_math.expand_matrix(mat, wires=self.wires, wire_order=wire_order)
+        return qp.math.expand_matrix(mat, wires=self.wires, wire_order=wire_order)
