+++ pennylane/resource/__init__.py
@@ -151,23 +151,3 @@ from .resource import (
 )
 from .expression import Expression
 from .specs import specs
-
-__all__ = [
-    "AlgorithmicError",
-    "ErrorOperation",
-    "SpectralNormError",
-    "algo_error",
-    "Resources",
-    "ResourcesOperation",
-    "SpecsResources",
-    "SymbolicSpecsResources",
-    "CircuitSpecs",
-    "add_in_series",
-    "add_in_parallel",
-    "mul_in_series",
-    "mul_in_parallel",
-    "resources_from_tape",
-    "substitute",
-    "Expression",
-    "specs",
-]
