+++ tests/ops/op_math/test_controlled2.py
@@ -13,19 +13,17 @@
 
 """Tests for the Controlled2 class."""
 
-import copy
-import pickle
 from typing import override
 
 import numpy as np
 import pytest
 
 import pennylane as qp
-from pennylane.ops.op_math.controlled import Controlled, ControlledOp
+from pennylane.ops.op_math.controlled import Controlled
 from pennylane.ops.op_math.controlled2 import Controlled2, ControlledOp2
 from pennylane.typing import Bool, Wire
 from pennylane.wires import Wires
-from tests.core.operator.operator2_utils import DynOp, OneWireDynOp
+from tests.core.operator.operator2_utils import OneWireDynOp
 
 # pylint: disable=unused-argument,too-few-public-methods
 
@@ -218,17 +216,7 @@ class TestControlled2:
         base = qp.H(0)
         op = ControlledOp2(base, control_wires=[1, 2])
         assert issubclass(ControlledOp2, Controlled)
-        assert issubclass(ControlledOp2, ControlledOp)
         assert isinstance(op, Controlled)
-        assert isinstance(op, ControlledOp)
-
-    @pytest.mark.parametrize("op_type", (qp.CNOT, qp.MultiControlledX, qp.Toffoli, qp.CZ))
-    def test_subclass_hook_does_not_match_specific_controlled_ops(self, op_type):
-        """Test that ControlledOp2 does not pretend to be a concrete controlled gate."""
-
-        op = ControlledOp2(qp.X(0), control_wires=[1])
-        assert not issubclass(ControlledOp2, op_type)
-        assert not isinstance(op, op_type)
 
     def test_simplify(self):
         """Tests the simplify method."""
@@ -430,13 +418,3 @@ class TestControlledOp2:
         op = OneWireDynOp(0.5, wires=[0])
         op = qp.ctrl(OneWireDynOp(0.5, wires=[0]), control=[1], control_values=0)
         assert isinstance(op, ControlledOp2)
-
-    @pytest.mark.parametrize(
-        "copy_fn", (copy.copy, copy.deepcopy, lambda obj: pickle.loads(pickle.dumps(obj)))
-    )
-    def test_copy_roundtrip(self, copy_fn):
-        """Test to make sure that copy roundtrips are sastisfied."""
-
-        op = ControlledOp2(DynOp(0.5, 0), control_wires=1)
-
-        assert op == copy_fn(op)
