+++ tests/ops/op_math/test_symbolic_op2_capture.py
@@ -28,7 +28,6 @@ pytestmark = [pytest.mark.jax, pytest.mark.capture]
 
 # pylint: disable=wrong-import-position
 from pennylane.capture.primitives import AbstractOperator, operator_p
-from tests.core.operator.operator2_utils import NonParametricOp
 
 
 def _single_op_eqn(jaxpr):
@@ -321,57 +320,6 @@ class TestNestedSymbolicOpCapture:
         inner_op = Adjoint2(RX2(0.7, wires=1)) if lazy else RX2(-0.7, wires=1)
         expected = ControlledOp2(inner_op, control_wires=[0])
         qp.assert_equal(op, expected, check_interface=False)
-
-
-@pytest.mark.parametrize("defined_outside", (True, False))
-def test_public_s_prod_binding(defined_outside):
-    """Tests that the public API for symbolic op captures properly."""
-
-    outside_op = NonParametricOp(0) if defined_outside else None
-
-    def f():
-        op = outside_op if defined_outside else NonParametricOp(0)
-        qp.s_prod(2.0, op)
-
-    cjaxpr = jax.make_jaxpr(f)()
-
-    eqns = cjaxpr.eqns
-
-    assert len(eqns) == 2  # operator and sprod
-    assert eqns[0].primitive.name == "operator"
-    assert eqns[0].params["op_cls"] is NonParametricOp
-
-    assert eqns[1].primitive.name == "SProd"
-
-    # SProd primitive consumes the op
-    assert eqns[0].outvars[0] == eqns[1].invars[1]
-
-
-def test_public_prod_binding():
-    """Tests that the public API for symbolic op captures properly."""
-
-    # NOTE: Have one op be outside trace context to
-    # cover the tracer-is-none fallback
-    outside_op = NonParametricOp(1)
-
-    def f():
-        qp.prod(outside_op, NonParametricOp(0))
-
-    cjaxpr = jax.make_jaxpr(f)()
-
-    eqns = cjaxpr.eqns
-
-    assert len(eqns) == 3  # op, op and sprod
-    assert eqns[0].primitive.name == "operator"
-    assert eqns[0].params["op_cls"] is NonParametricOp
-    assert eqns[1].primitive.name == "operator"
-    assert eqns[1].params["op_cls"] is NonParametricOp
-
-    assert eqns[2].primitive.name == "Prod"
-
-    # Prod primitive consumes the ops
-    assert eqns[1].outvars[0] == eqns[2].invars[0]
-    assert eqns[0].outvars[0] == eqns[2].invars[1]
 
 
 if __name__ == "__main__":
